feat: replace temporary menu page
